requirements.txt
path.py==8.1.1
Django==1.8.4
pelican==3.6.3
docutils==0.12
django-compressor==1.5
django-libsass==0.4
django-downloadview==1.8
unicode-slugify==0.1.3